New SCALP script with configs and unmount fix
Unmounting while program is running has been fixed by adding an unmount rule in enamel_usb_autostart.rules (triggered when a device is removed), which simply runs unmountDrive.sh (which just calls umount on the specified drive).

SCALP.sh has been modified with extra variables, which are grabbed from config.txt. Depending on the state of the variables, it acts accordingly (for example, if onefileautoplay is false, will copy over the file and then play).

changeconfig.sh has been added. Java calls changeconfig.sh with arguments that specify which variable to change, and to what value to change it to. changeconfig.sh uses "sed" to find the specified variable in config.txt, and changes the value. 
For example, it finds the text "onefileautoplay" in config.txt and replaces the *whole line* with 
"onefileautoplay=(changed value)".