Yassin's first commit; fixed a couple of bugs with the Authoring App.
This commit is from the Macbook testing laptop, so we get two birds with
one stone.
Fixed the bug where the Authoring App closed after testing.
Fixed a bug where scenarios were altered upon saving them.

I also realized that the TestScenario button saves the currently running
scenario first then plays it. Meaning if you were running the app and
building a scenario, then you save and exit, then you open the app and
click the test scenario button and navigate to your previously created
one the app would override your own scenario with an empty one and play
that. This is basically because the TestScenario button is a modified
Export button, and so it exports whatever scenario is being build right
now and tests it. As it stands you can work around that just by pressing
the "Import" button before the "Test Scenario" button, but I imagine we
can fix this just by replacing the "Export" part of the Test Scenario
code with Import button code.