Implemented direct item add hotkeys, confirm window on exit + many bug fixes and minor changes
Renamed Import to Load and Export to Save.
Authoring App now confirms if you want to close, and warns you that
unsaved changes will be lost, this is to prevent accidental exits that
would result in lost data.
Implemented hotkeys that allow the user to directly add "items" to the
scenario, all mapped to ALT. The specific mappings we can change but the
framework is there now. 
Every item type in NewButtonListener didn't verify if the user canceled
out without entering any values, so I added if statements to most.